% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nigeria.R
\name{choropleth_ng}
\alias{choropleth_ng}
\title{Choropleth Maps of Nigeria}
\usage{
choropleth_ng(x, breaks, base.col = c("red", "grey", "green", "blue"),
  ...)
}
\arguments{
\item{x}{A variable to be mapped}

\item{breaks}{Numeric vector > 3 and < 10 representing the limits of the
categories for variable \code{x}.}

\item{base.col}{The base colour upon which the colour scale is based}

\item{...}{Arguments to be passed to \code{\link[maps]{map}}}
}
\description{
Choropleth Maps of Nigeria
}
