% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nigeria.R
\name{cmap}
\alias{cmap}
\title{Choropleth Maps of Nigeria}
\usage{
cmap(x, base.col, breaks, ...)
}
\arguments{
\item{x}{A variable to be mapped}

\item{base.col}{The base colour upon which the colour scheme is based}

\item{breaks}{Numeric vector > 3 and < 10 representing the limits of the
categories for variable \code{x}.}

\item{...}{Arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Choropleth Maps of Nigeria
}
