% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{fix_mobile}
\alias{fix_mobile}
\title{Fix up mobile numbers}
\usage{
fix_mobile(column)
}
\arguments{
\item{column}{A character vector of numerical string characters.}
}
\value{
An update vector, usually the column of a data frame.
}
\description{
Fixes up mobile numbers to a uniform text format.
}
\details{
This format is specific to that used in a given location - for
now the function is useful only for Nigeria mobile numbers which come in
the format expressed by the regex pattern \code{"^0[7-9][0-1][0-9]{8}$"}.
}
