% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/political.R
\name{is_state}
\alias{is_state}
\title{Test an Object for States}
\usage{
is_state(x, test = c("all", "selected"))
}
\arguments{
\item{x}{A character vector to be tested.}

\item{test}{The type of test to be carried out - on the vector as a whole 
i.e. \code{all} (the default argument) or on the individual elements i.e.
\code{selected}.}
}
\value{
A logical vector.
}
\description{
Test an Object for States
}
