% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/political.R
\name{lgalist}
\alias{lgalist}
\title{Names of Local Government Areas in Nigeria}
\usage{
lgalist(state = NULL)
}
\arguments{
\item{state}{State of the Federation}
}
\value{
Local Government Areas

A character vector with the list of Local Government Areas.
}
\description{
Names of Local Government Areas in Nigeria
}
\details{
When a State is chosen by passing in `state`, the function obtains
list of Local Government Areas in that particular State.
}
\examples{
lgalist("Abia")
}

