% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naijR-package.R
\docType{package}
\name{naijR-package}
\alias{naijR}
\alias{naijR-package}
\title{naijR: Operations to Ease Data Analyses Specific to Nigeria}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of convenience functions as well as geographical/political data about Nigeria, aimed at simplifying work with data and information that are specific to the country.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/naijR/}
  \item Report bugs at \url{https://github.com/ropensci/naijR/issues}
}

}
\author{
\strong{Maintainer}: Victor Ordu \email{victorordu@outlook.com} (\href{https://orcid.org/0000-0003-3716-0668}{ORCID})

Other contributors:
\itemize{
  \item Laura DeCicco \email{ldecicco@usgs.gov} [contributor]
  \item Emily Riederer \email{emilyriederer@gmail.com} [reviewer]
  \item Margaret Siple [reviewer]
  \item Cagri Alican [reviewer]
}

}
\keyword{internal}
