% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-regions.R
\name{print.regions}
\alias{print.regions}
\title{Print regions}
\usage{
\method{print}{regions}(x, ...)
}
\arguments{
\item{x}{An object of class \code{regions}}

\item{...}{Additional arguments, though not set. Left for future use}
}
\description{
Print regions
}
