% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regdata.R
\docType{data}
\name{states_nigeria}
\alias{states_nigeria}
\title{States of Nigeria}
\format{
A \code{data.frame} with 37 rows and 2 columns
\describe{
  \item{isocode}{ISO 3661 Alpha 2 code}
  \item{state}{Name of the State or Territory}
  \item{gpz}{Geo-political zone}
}
}
\usage{
data(states_nigeria)
}
\description{
A dataset of the 36 States and Federal Capital Territory of Nigeria.
}
\keyword{datasets}
