% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ward_data.R
\docType{data}
\name{ward_data}
\alias{ward_data}
\title{Nigeria political wards by Local Government Area}
\format{A data frame with 8799 rows and 2 columns:
\describe{
   \item{LGA}{Name of Local Government Area}
   \item{Ward}{Name of political ward}
}}
\source{
\url{http://www.inecnigeria.org/wp-content/uploads/2013/03/RA-LGA-ANALYSIS-NATIONWIDE.pdf}
}
\usage{
ward_data
}
\description{
A complete list of political wards in the Local Government Areas of Nigeria.
}
\keyword{datasets}

