% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixmob.R
\name{fix_mobile}
\alias{fix_mobile}
\title{Fix mobile numbers}
\usage{
fix_mobile(x)
}
\arguments{
\item{x}{Values to be fixed, represented either as \code{numeric} or
\code{character} vectors.}
}
\value{
The updated vector, usually the column of a data frame.
}
\description{
Fixes up local mobile phone numbers to a uniform text format.
}
\details{
This format is specific to that used in a given location - for
now the function is useful only for Nigeria mobile numbers, which come in
the format expressed by the regex pattern \code{"^0[7-9][0-1][0-9]{8}$"}.
}
\note{
There is an option for producing warnings on any mobile number entries
that may have been removed from the vector, by setting the option
\code{verbose} to \code{TRUE}.
}
\examples{
fix_mobile("803-123-4567")    # Adds leading '0" and removes separators

}
