% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{fix_region}
\alias{fix_region}
\title{Fix Region Names}
\usage{
fix_region(x, ...)
}
\arguments{
\item{x}{An S3 object of class \code{states} or \code{lgas}}

\item{...}{This argument is placed for possible use in the near future.}
}
\value{
The transformed object. If all names are correct, the object is
returned unchanged.
}
\description{
Correct any misspelt names of administrative regions i.e. States and LGAs
}
\details{
The function will look through a character vector and try to 
determine if State names have been wrongly entered. This presupposes that
the atomic vector is of type \code{character}. It does not test any missing
values in the vector, leaving them untouched.
}
\note{
An updated version would include the ability to adjust the 
\href{https://en.wikipedia.org/wiki/Levenshtein_distance}{Levenshtein 
distance}, which will empower users to tune the function's sensitivity.
}
