% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixreg.R
\name{fix_region_manual}
\alias{fix_region_manual}
\title{Fix Spelling of Regions Manually}
\usage{
fix_region_manual(x, wrong, correct)
}
\arguments{
\item{x}{The object to be modified}

\item{wrong}{The misspelt element(s) of \code{x}.}

\item{correct}{The correction that is to be applied to the misspelt element(s)}
}
\description{
Enable users to interactively and directly change to spelling of States
and/or Local Government Areas (LGAs)
}
