% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isreg.R
\name{is_lga}
\alias{is_lga}
\title{Test for Local Government Areas}
\usage{
is_lga(x)
}
\arguments{
\item{x}{An object of type \code{character}. This includes higher
dimension object classes like \code{matrix} and \code{array}.}
}
\value{
A logical vector the same length as the input object. Each
element that is not a valid Local Government Area will evaluate to
\code{FALSE}.
}
\description{
Checks a given object for Local Government Areas, represented as
strings.
}
\examples{
is_lga(c("Pankshen", "Pankshin"))

}
