% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgas_nigeria.R
\name{lgas_ng}
\alias{lgas_ng}
\title{List Local Government Areas}
\usage{
lgas_ng(ng.state = NULL)
}
\arguments{
\item{ng.state}{Character vector of length 1, whose value is the name of a
State in the Federation of Nigeria. Default is \code{NULL}, in which case
all the countries Local Government Areas will be returned.}
}
\value{
A character vector containing the names of Local Government Areas.
}
\description{
List Local Government Areas
}
