% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgas_nigeria.R
\docType{data}
\name{lgas_nigeria}
\alias{lgas_nigeria}
\title{Local Government Areas of Nigeria}
\format{A dataframe with 774 rows and 2 columns
\describe{
   \item{lga}{Local Government Area}
   \item{state}{State of the Federation}
   }}
\usage{
lgas_nigeria
}
\description{
A dataset containing the 774 Local Government Areas of Nigeria
}
\keyword{datasets}
