% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nigeria.R
\name{map_ng}
\alias{map_ng}
\title{Map of Nigeria}
\usage{
map_ng(fill = TRUE, col = "lightgrey", adjoin = FALSE)
}
\arguments{
\item{fill}{Whether to fill the plot or not}

\item{col}{The colour of the filled portion of the plot.}

\item{adjoin}{Whether to display borders of neighbouring countries.}
}
\description{
Map of Nigeria
}

