% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_ng}
\alias{map_ng}
\title{Map of Nigeria}
\usage{
map_ng(
  region = character(),
  data = NULL,
  x = NULL,
  y = NULL,
  breaks = NULL,
  categories = NULL,
  excluded = NULL,
  exclude.fill = NULL,
  title = NULL,
  caption = NULL,
  show.neighbours = FALSE,
  show.text = FALSE,
  legend.text = NULL,
  leg.x = deprecated(),
  leg.y = deprecated(),
  leg.title,
  plot = TRUE,
  leg.orient = deprecated(),
  ...
)
}
\arguments{
\item{region}{A character vector of regions to be displayed. This could be 
States or Local Government Areas.}

\item{data}{An object containing data, principally the variables required to
plot in a map.}

\item{x, y}{Numeric object or factor (or coercible to one). See \emph{Details}.}

\item{breaks}{Numeric. A vector of length >= 1. If a single value i.e.
scalar, it denotes the expected number of breaks. Internally, the function
will attempt to compute appropriate category sizes or fail if out-of bounds. 
Where length is >= 3L, it is expected to be an arithmetic sequence that 
represents category bounds as for \code{\link[base]{cut}} (applicable 
only to choropleth maps).}

\item{categories}{The legend for the choropleth-plotted categories. If not 
defined, internally created labels are used.}

\item{excluded}{Regions to be excluded from a choropleth map.}

\item{exclude.fill}{Colour-shading to be used to indicate \code{excluded}
regions. Must be a vector of the same length as \code{excluded}.}

\item{title, caption}{An optional string for annotating the map.}

\item{show.neighbours}{Logical; \code{TRUE} to display the immediate vicinity
neighbouring regions/countries.}

\item{show.text}{Logical. Whether to display the labels of regions.}

\item{legend.text}{Logical (whether to show the legend) or character vector
(actual strings for the legend). The latter will override whatever is 
provided by \code{categories}, giving the user additional control.}

\item{leg.x, leg.y}{Numeric. Position of the legend (deprecated).}

\item{leg.title}{String. The legend title. If missing, a default value is
acquired from the data. To turn off the legend title, pass \code{NULL}.}

\item{plot}{Logical. Turn actual plotting of the map off or on.}

\item{leg.orient}{The orientation of the legend i.e. whether horizontal or
vertical (deprecated).}

\item{...}{Further arguments passed to \code{\link[sf]{plot}}}
}
\value{
An object of class \code{sf}, which is a standard format containing 
the data used to draw the map and thus can be used by this and other 
popular R packages to visualize the spatial data.
}
\description{
Maps of the Federal Republic of Nigeria that are based on the basic
plotting idiom utilised by \link[maps:map]{maps:map} and its variants.
}
\details{
The default value for \code{region} is to print all State boundaries.
\code{data} enables the extraction of data for plotting from an object
of class \code{data.frame}. Columns containing regions (i.e. States as well as
supported sub-national jurisdictions) are identified. The argument also
provides context for quasiquotation when providing the \code{x} and
\code{y} arguments.

For \code{x} and \code{y}, when both arguments are supplied, they are taken
to be point coordinates, where \code{x} represent longitude and \code{y}
latitude. If only \code{x} is supplied, it is assumed that the intention of
the user is to make a choropleth map, and thus, numeric vector arguments are
converted into factors i.e. number classes. Otherwise factors or any object 
that can be coerced to a factor should be used.

For plain plots, the \code{col} argument works the same as with
\code{\link[maps]{map}}. For choropleth maps, the colour provided represents 
a (sequential) colour palette based on \code{RColorBrewer::brewer.pal}. The 
available colour options can be checked with 
\code{getOption("choropleth.colours")} and this can also be modified by the 
user.

If the default legend is unsatisfactory, it is recommended that the user
sets the \code{legend.text} argument to \code{FALSE}; the next function
call should be \code{\link[graphics]{legend}} which will enable finer
control over the legend.
}
\note{
When adjusting the default colour choices for choropleth maps, it is
advisable to use one of the sequential palettes. For a list of of available
palettes, especially for more advanced use, review 
\code{RColorBrewer::display.brewer.all}.
}
\examples{
\dontrun{
map_ng() # Draw a map with default settings
map_ng(states("sw"))
map_ng("Kano")
}

}
