% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nigeria.R
\name{map_ng}
\alias{map_ng}
\title{Map of Nigeria}
\usage{
map_ng(show.neighbours = FALSE, mar = c(2.1, 2.1, par("mar")[3], 0.1),
  ...)
}
\arguments{
\item{show.neighbours}{logical; \code{TRUE} to display borders of
neighbouring countries.}

\item{mar}{Plot margins as in \code{\link[graphics]{par}}}

\item{...}{Further arguments for function \code{\link[maps]{map}}}
}
\value{
An object of class \code{map}, invisibly; as a side-effect,
results in the drawing of a map of Nigeria.

An object of class \code{maps} containing the data used to draw the
map and which can be used for additional calls to \code{\link[maps]{map}}.
}
\description{
A basic map of the Federal Republic of Nigeria.
}
\details{
This function is essentially a wrapper to \code{maps::map}.
}
\examples{
map_ng() # Draw a map with default settings

}
