% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nigeria.R
\name{map_ng}
\alias{map_ng}
\title{Map of Nigeria}
\usage{
map_ng(show.neighbours = FALSE, ...)
}
\arguments{
\item{show.neighbours}{logical; \code{TRUE} to display borders of
neighbouring countries.}

\item{...}{Further arguments for function \code{\link[maps]{map}}}
}
\value{
\code{NULL} invisibly; results in the drawing of a map of Nigeria.
}
\description{
A basic map of the Federal Republic of Nigeria.
}
\details{
This function is essentially a wrapper to \code{maps::map}.
}
\examples{
map_ng() # Draw a map with default settings
map_ng(TRUE) # Display portions of neighbouring countries' borders

}
