% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nigeria.R
\name{map_ng}
\alias{map_ng}
\title{Map of Nigeria}
\usage{
map_ng(state = character(), flavour = c("plain", "choropleth"),
  data = NULL, region = NULL, value = NULL, breaks = NULL,
  show.neighbours = FALSE, col = if (flavour == "plain") 1L else NULL,
  fill = FALSE, label = FALSE, ...)
}
\arguments{
\item{state}{A character vector of a list of Nigerian States to be displayed.
The default value is to print all States.}

\item{flavour}{The type of map to be drawn. Current options are \code{plain}
and \code{choropleth}.}

\item{data}{An object containing data, principally the variables required to
plotted in a map (applicable only to \code{choropleth})}

\item{region}{The subdivision for choropleth mapping.}

\item{value}{The value to be categorised for choropleth mapping.}

\item{breaks}{Categories to be plotted on a map (applicable only 
to \code{choropleth})}

\item{show.neighbours}{logical; \code{TRUE} to display borders of
neighbouring countries.}

\item{col}{A character vector of R \code{\link[grDevices]{colours()}} to be
used in the map. If \code{fill == TRUE}, colours are applied to the mapped
area; otherwise it is applied to the borders.}

\item{fill}{Logical. Whether to fill the mapped area or not.}

\item{label}{Logical. Apply labels to the map.}

\item{...}{Further arguments for function \code{\link[maps]{map}}}
}
\value{
An object of class \code{map}, invisibly; as a side-effect,
results in the drawing of a map of Nigeria.

An object of class \code{maps} containing the data used to draw the
map and which can be used for additional calls to \code{\link[maps]{map}} or
other similar functions (e.g. \code{\link[graphics]{plot}}).
}
\description{
A map of the Federal Republic of Nigeria.
}
\details{
This function is essentially a wrapper to \code{maps::map}.

...
}
\examples{
map_ng() # Draw a map with default settings

## Choropleth
set.seed(4)
vals <- sample(0:6, 37, TRUE)
brk <- seq(0, 6, 2)
map_ng(flavour = 'choropleth', var = vals, breaks = brk)

}
