% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naijR-package.R
\docType{package}
\name{naijR-package}
\alias{naijR}
\alias{naijR-package}
\title{naijR: Operations to Ease Data Analyses Specific to Nigeria}
\description{
The naijR package is essentially an R package about Nigeria and for Nigeria.
}
\details{
naijR contains a number of functions that facilitate the
management of data sets of interest including data cleaning and wrangling,
as well as making available a number of facilities for geospatial data
visualisation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/naijR/}
  \item Report bugs at \url{https://github.com/ropensci/naijR/issues}
}

}
\author{
Victor Ordu
}
\keyword{internal}
