% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regdata.R
\docType{data}
\name{ngdist}
\alias{ngdist}
\title{Distance By Road of Nigerian State Capitals}
\format{
An object of class \code{dist}.
}
\usage{
ngdist
}
\description{
A distance matrix with the road distance in kilometers for Nigeria's
State capitals.
}
\keyword{datasets}
