% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/political.R
\name{states}
\alias{states}
\title{Display States of the Federal Republic of Nigeria}
\usage{
states(gpz = NULL, sorted = TRUE, full.names = FALSE)
}
\arguments{
\item{gpz}{Geopolitical zone. Default is \code{NULL}; optionally \code{"nc",
"ne", "nw", "se", "ss"} and \code{"sw"} (see \code{Details}).}

\item{sorted}{Whether the results to be all sorted in increasing 
alphabetical order (default value is \code{TRUE}).}

\item{full.names}{the complete appellation for the States}
}
\value{
The States of Nigeria as a whole or by zones, as a character vector
}
\description{
Display States of the Federal Republic of Nigeria
}
\details{
gpz A geo-political zone, in the Nigerian 
context, is a national subdivision that groups contiguous states. 
Historically, they arise from subnational administrative divisions 
known as 'regions' that existed at the time of the country's independence.
There are 6 zones - North-Central, North-East, North-West, South-East,
South-South and South-West.
}
\examples{
states()  # lists names of all States
states("se")  # lists States in South-East zone
}
