% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{states}
\alias{states}
\title{Display States of the Federal Republic of Nigeria}
\usage{
states(gpz = NULL)
}
\arguments{
\item{gpz}{Geopolitical zone}
}
\value{
The States of Nigeria as a whole or by zones
}
\description{
Display States of the Federal Republic of Nigeria
}
\examples{
states()
states(2)
states("se")
}

