% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regiondata.R
\docType{data}
\name{states_nigeria}
\alias{states_nigeria}
\title{States of Nigeria}
\format{
A \code{data.frame} with 37 rows and 2 columns
\describe{
  \item{Code}{ISO 3661 Alpha 2 code}
  \item{Name}{Name of the State or Territory}
}
}
\usage{
states_nigeria
}
\description{
A dataset of the 36 States and Federal Capital Territory of Nigeia
}
\keyword{datasets}
